+++ src/app/providers/index.ts
@@ -1,5 +1,5 @@
-import { withErrorBoundary } from "@/app/providers/withErrorBoundary";
-import { withErrorToast } from "@/app/providers/withErrorToast";
+import { withErrorBoundary } from "@/shared/hoc/with-error-boundary";
+import { withErrorToast } from "@/shared/hoc/with-error-toast";
 import { compose } from '@/shared/lib';
 
 export const withProviders = compose(